feat(scan): Move shortDateTime to @votingworks/utils
