+++ libs/utils/src/date.ts
@@ -81,3 +81,14 @@ export function getDaysInMonth(year: number, month: number): DateTime[] {
   }
   return days;
 }
+
+function z2(number: number) {
+  return number.toString().padStart(2, '0');
+}
+
+export function shortDateTime(iso8601Timestamp: string) {
+  const d = new Date(iso8601Timestamp);
+  return `${d.getFullYear()}-${z2(d.getMonth() + 1)}-${z2(
+    d.getDate()
+  )} ${d.getHours()}:${z2(d.getMinutes())}:${z2(d.getSeconds())}`;
+}
